+++ VKApp/VKApp/Controllers/BigPhotosUserViewController.swift
@@ -18,7 +18,6 @@ final class BigPhotosUserViewController: UIViewController {
 
     private var userPhotoNames: [String] = []
     private var currentUserPhotoIndex = 0
-    private let vkNetworkService = VKNetworkService()
 
     // MARK: - Lifecycle
 
@@ -77,17 +76,11 @@ final class BigPhotosUserViewController: UIViewController {
         currentUserPhotoTrailingConstraint.constant = 0
         currentUserPhotoLeadingConstraint.constant = 0
         currentUserPhotoImageView.layer.zPosition = 1
-        vkNetworkService.setupImage(
-            urlPath: userPhotoNames[currentUserPhotoIndex],
-            imageView: currentUserPhotoImageView
-        )
+        currentUserPhotoImageView.setupImage(urlPath: userPhotoNames[currentUserPhotoIndex])
         nextUserPhotoTrailingConstraint.constant = -view.frame.width
         nextUserPhotoLeadingConstraint.constant = view.frame.width
         nextUserPhotoImageView.layer.zPosition = 2
-        vkNetworkService.setupImage(
-            urlPath: userPhotoNames[currentUserPhotoIndex + 1],
-            imageView: nextUserPhotoImageView
-        )
+        nextUserPhotoImageView.setupImage(urlPath: userPhotoNames[currentUserPhotoIndex + 1])
         view.layoutIfNeeded()
         currentUserPhotoIndex += 1
     }
@@ -113,17 +106,11 @@ final class BigPhotosUserViewController: UIViewController {
         currentUserPhotoTrailingConstraint.constant = 0
         currentUserPhotoLeadingConstraint.constant = 0
         currentUserPhotoImageView.layer.zPosition = 2
-        vkNetworkService.setupImage(
-            urlPath: userPhotoNames[currentUserPhotoIndex],
-            imageView: currentUserPhotoImageView
-        )
+        currentUserPhotoImageView.setupImage(urlPath: userPhotoNames[currentUserPhotoIndex])
         nextUserPhotoTrailingConstraint.constant = 50
         nextUserPhotoLeadingConstraint.constant = 50
         nextUserPhotoImageView.layer.zPosition = 1
-        vkNetworkService.setupImage(
-            urlPath: userPhotoNames[currentUserPhotoIndex - 1],
-            imageView: nextUserPhotoImageView
-        )
+        nextUserPhotoImageView.setupImage(urlPath: userPhotoNames[currentUserPhotoIndex - 1])
         view.layoutIfNeeded()
         currentUserPhotoIndex -= 1
     }
@@ -135,10 +122,7 @@ final class BigPhotosUserViewController: UIViewController {
 
     private func setupImageViews() {
         guard 0 ..< userPhotoNames.count ~= currentUserPhotoIndex else { return }
-        vkNetworkService.setupImage(
-            urlPath: userPhotoNames[currentUserPhotoIndex],
-            imageView: currentUserPhotoImageView
-        )
+        currentUserPhotoImageView.setupImage(urlPath: userPhotoNames[currentUserPhotoIndex])
     }
 
     private func addSwipeToView() {
