+++ VKApp/VKApp/Extensions/UIImageViewExtensions.swift
@@ -1,0 +1,21 @@
+// UIImageViewExtensions.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+import UIKit
+
+extension UIImageView {
+    func setupImage(urlPath: String?) {
+        guard
+            let urlPath = urlPath,
+            let url = URL(string: urlPath)
+        else { return }
+        DispatchQueue.global().async {
+            let data = try? Data(contentsOf: url)
+            DispatchQueue.main.async {
+                guard let data = data else { return }
+                self.image = UIImage(data: data)
+            }
+        }
+    }
+}
