+++ VKApp/VKApp/Services/VKNetworkService.swift
@@ -73,9 +73,8 @@ final class VKNetworkService {
     func fetchFriendsVK(completion: @escaping ([ItemPerson]) -> Void) {
         let path = Constants.methodText + Constants.friendsGetText
         let url = "\(Constants.baseUrl)\(path)"
-        Alamofire.request(url, method: .get, parameters: parametersFriendsVK).responseData { [weak self] response in
+        Alamofire.request(url, method: .get, parameters: parametersFriendsVK).responseData { response in
             guard
-                let self = self,
                 let data = response.value,
                 let items = try? JSONDecoder().decode(Person.self, from: data).response.items
             else { return }
@@ -85,13 +84,12 @@ final class VKNetworkService {
     }
 
     func fetchPhotosVK(userID: String, completion: @escaping ([String]) -> Void) {
-        let path = Constants.methodText + Constants.photosGetAllText
+        let path = "\(Constants.methodText)\(Constants.photosGetAllText)"
         let url = "\(Constants.baseUrl)\(path)"
         var parametersPhotos = generalParameters
         parametersPhotos[Constants.ownerIdText] = userID
-        Alamofire.request(url, method: .get, parameters: parametersPhotos).responseData { [weak self] response in
+        Alamofire.request(url, method: .get, parameters: parametersPhotos).responseData { response in
             guard
-                let self = self,
                 let data = response.value,
                 let items = try? JSONDecoder().decode(Photo.self, from: data).response.items
             else { return }
@@ -105,11 +103,10 @@ final class VKNetworkService {
     }
 
     func fetchUserGroupsVK(completion: @escaping ([VKGroups]) -> Void) {
-        let path = Constants.methodText + Constants.groupsGetText
+        let path = "\(Constants.methodText)\(Constants.groupsGetText)"
         let url = "\(Constants.baseUrl)\(path)"
-        Alamofire.request(url, method: .get, parameters: parametersGroupVK).responseData { [weak self] response in
+        Alamofire.request(url, method: .get, parameters: parametersGroupVK).responseData { response in
             guard
-                let self = self,
                 let data = response.value,
                 let items = try? JSONDecoder().decode(VKGroup.self, from: data).response.items
             else { return }
@@ -119,7 +116,7 @@ final class VKNetworkService {
     }
 
     func fetchSearchGroupsVK(searchText: String, completion: @escaping ([VKGroups]) -> Void) {
-        let path = Constants.methodText + Constants.groupsSearchText
+        let path = "\(Constants.methodText)\(Constants.groupsSearchText)"
         let url = "\(Constants.baseUrl)\(path)"
         var parametersSearchGroupVK = generalParameters
         parametersSearchGroupVK[Constants.qText] = searchText
@@ -146,19 +143,5 @@ final class VKNetworkService {
             URLQueryItem(name: Constants.vText, value: Constants.vValueText)
         ]
         return urlComponents
-    }
-
-    func setupImage(urlPath: String?, imageView: UIImageView) {
-        guard
-            let urlPath = urlPath,
-            let url = URL(string: urlPath)
-        else { return }
-        DispatchQueue.global().async {
-            let data = try? Data(contentsOf: url)
-            DispatchQueue.main.async {
-                guard let data = data else { return }
-                imageView.image = UIImage(data: data)
-            }
-        }
     }
 }
