+++ VKApp/VKApp/Views/FriendsUserTableViewCell.swift
@@ -18,10 +18,6 @@ final class FriendsUserTableViewCell: UITableViewCell {
     @IBOutlet private var friendPhotoImageView: UIImageView!
     @IBOutlet private var shadowView: ShadowView!
 
-    // MARK: - Private Properties
-
-    private let vkNetworkService = VKNetworkService()
-
     // MARK: - Lifecycle
 
     override func awakeFromNib() {
@@ -42,7 +38,7 @@ final class FriendsUserTableViewCell: UITableViewCell {
 
     func configure(user: User) {
         friendNameLabel.text = user.userName
-        vkNetworkService.setupImage(urlPath: user.userPhotoURLText, imageView: friendPhotoImageView)
+        friendPhotoImageView.setupImage(urlPath: user.userPhotoURLText)
         self.user = user
     }
