+++ VKApp/VKApp/Views/SearchGroupTableViewCell.swift
@@ -20,16 +20,12 @@ final class SearchGroupTableViewCell: UITableViewCell {
 
     var group = Group(groupName: Constants.emptyText, groupPhotoName: Constants.emptyText)
 
-    // MARK: - Private Properties
-
-    private let vkNetworkService = VKNetworkService()
-
     // MARK: - Public Methods
 
     func configure(group: Group) {
         selectionStyle = .none
         groupNameLabel.text = group.groupName
-        vkNetworkService.setupImage(urlPath: group.groupPhotoName, imageView: groupPhotoImageView)
+        groupPhotoImageView.setupImage(urlPath: group.groupPhotoName)
         self.group = group
     }
 }
